+++ test/Transport/SendmailTest.php
@@ -10,6 +10,7 @@
 namespace ZendTest\Mail\Transport;
 
 use ReflectionMethod;
+use ReflectionProperty;
 use Zend\Mail\Address\AddressInterface;
 use Zend\Mail\AddressList;
 use Zend\Mail\Header;
@@ -19,7 +20,6 @@ use Zend\Mail\Transport\Exception\RuntimeException;
 use Zend\Mail\Transport\Sendmail;
 
 /**
- * @group      Zend_Mail
  * @covers Zend\Mail\Transport\Sendmail<extended>
  */
 class SendmailTest extends \PHPUnit_Framework_TestCase
@@ -211,5 +211,15 @@ class SendmailTest extends \PHPUnit_Framework_TestCase
 
         $parameters = $r->invoke($this->transport, $message->reveal());
         $this->assertEquals(' -f' . escapeshellarg($injectedEmail), $parameters);
+    }
+
+    public function testTrimmedParameters()
+    {
+        $this->transport->setParameters([' -R', 'hdrs ']);
+
+        $r = new ReflectionProperty($this->transport, 'parameters');
+        $r->setAccessible(true);
+
+        $this->assertSame('-R hdrs', $r->getValue($this->transport));
     }
 }